Struct google_urlshortener1::Url

pub struct Url {
    pub status: Option<String>,
    pub analytics: Option<AnalyticsSummary>,
    pub kind: Option<String>,
    pub created: Option<String>,
    pub long_url: Option<String>,
    pub id: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert url (request|response)

get url (response)

Fields

status

Status of the target URL. Possible values: "OK", "MALWARE", "PHISHING", or "REMOVED". A URL might be marked "REMOVED" if it was flagged as spam, for example.

analytics

A summary of the click analytics for the short and long URL. Might not be present if not requested or currently unavailable.

kind

The fixed string "urlshortener#url".

created

Time the short URL was created; ISO 8601 representation using the yyyy-MM-dd'T'HH:mm:ss.SSSZZ format, e.g. "2010-10-14T19:01:24.944+00:00".

long_url

Long URL, e.g. "http://www.google.com/". Might not be present if the status is "REMOVED".

id

Short URL, e.g. "http://goo.gl/l6MS".